requirements.txt
# Post Veille IA - Dépendances Python
# Sprint 1

# === COLLECTEURS ===

# RSS
feedparser>=6.0.10

# HTTP requests (Jina, Reddit)
requests>=2.31.0

# Configuration
pyyaml>=6.0.1

# === BASE DE DONNÉES ===
# SQLite intégré à Python, pas de dépendance externe

# === UTILITAIRES ===

# Variables d'environnement
python-dotenv>=1.0.0

# Dates
python-dateutil>=2.8.2

# === SPRINT 2 ===

# YouTube transcripts
youtube-transcript-api>=1.2.0

# === DISCORD BOT ===

# Discord API (bot officiel, pas selfbot)
discord.py>=2.3.0

# Async SQLite pour le bot
aiosqlite>=0.19.0

# HTTP async (webhooks)
aiohttp>=3.9.0

# === OPTIONNEL ===

# Logging amélioré
# rich>=13.0.0
